+++ src/core/normalizer.js
@@ -1,11 +1,7 @@
-import { readFileSync, writeFileSync } from 'fs';
+import { readFileSync } from 'fs';
 import { load } from 'js-yaml';
 import { resolve } from 'path';
-import sh from 'shelljs';
-import { Providers } from '../utils/data-schema.js';
 import { SpatialEntity } from '../utils/spatial-schema.js';
-import { importCsv } from './csv-normalizer.js';
-import { importFromList } from './import-list-normalizer.js';
 
 
 /** The default order that properties should show in objects */
@@ -20,50 +16,6 @@ const DEFAULT_PROPERTY_ORDER = [
   'section_size',
   'rui_location',
 ];
-
-/**
- * This function normalizes the registration data from a YAML file to a JSON-LD format and writes it to a file as output.
- *  @param { string } context - The directory path of registration.yaml file.
- */
-export async function normalizeRegistrations(context) {
-  const ruiLocationsDir = resolve(context.doPath, 'registrations');
-  const data = loadFile(context.doPath, 'registrations.yaml', Providers);
-  // Normal Normalization
-  let normalized = '';
-  for (const provider of data) {
-    if (provider.provider_name) {
-      normalized = await normalizeRegistration(data, ruiLocationsDir);
-    }
-  }
-  //CSV Normalization
-  let csv_normalized = '';
-  for (const csv of data) {
-    if (csv.import_from_csv) {
-      csv_normalized = await importCsv(csv.import_from_csv, csv.fields, csv.baseIri);
-    }
-  }
-
-  //Importing from list (web or local) files using filters
-  let import_list_normalized = '';
-  for (const import_list of data) {
-    if (import_list.imports) {
-      import_list_normalized = await importFromList(import_list.imports, import_list.filter);
-    }
-  }
-
-
-  const final = convertToJsonLd(normalized, csv_normalized, import_list_normalized);
-
-  const ruiLocationsOutputPath = resolve(
-    context.doPath,
-    'rui_locations.jsonld'
-  );
-  writeFileSync(ruiLocationsOutputPath, JSON.stringify(final, null, 2));
-  sh.cp(
-    resolve(context.processorHome, 'src/ccf-eui-template.html'),
-    resolve(context.doPath, 'index.html')
-  );
-}
 
 /**
  * This function ensures the registration.yaml file has id, label, description, and link at Provider, Donor, Section, Block and Dataset levels.
@@ -184,7 +136,7 @@ function loadFile(dir, file, schema) {
  * @param { object[] } container
  * @param { string[] } propOrder
  */
-function ensurePropertyOrder(
+export function ensurePropertyOrder(
   obj,
   container,
   propOrder = DEFAULT_PROPERTY_ORDER
@@ -459,84 +411,4 @@ function* enumerate(arr) {
   for (let i = 0; i < (arr ?? []).length; i++) {
     yield [arr[i], i];
   }
-}
-
-/**
- * This function is used to convert the above generated schema to JsonLd format.
- * @param { object } data - The data, which was ensured above, and which needs to be converted to JsonLd format.
- */
-export function convertToJsonLd(normalized, csv_normalized, import_list_normalized) {
-  const data = {
-    '@context': {
-      '@base': 'http://purl.org/ccf/latest/ccf-entity.owl#',
-      '@vocab': 'http://purl.org/ccf/latest/ccf-entity.owl#',
-      ccf: 'http://purl.org/ccf/latest/ccf.owl#',
-      rdfs: 'http://www.w3.org/2000/01/rdf-schema#',
-      label: 'rdfs:label',
-      description: 'rdfs:comment',
-      link: {
-        '@id': 'rdfs:seeAlso',
-        '@type': '@id',
-      },
-      samples: {
-        '@reverse': 'has_donor',
-      },
-      sections: {
-        '@id': 'has_tissue_section',
-        '@type': '@id',
-      },
-      datasets: {
-        '@id': 'has_dataset',
-        '@type': '@id',
-      },
-      rui_location: {
-        '@id': 'has_spatial_entity',
-        '@type': '@id',
-      },
-      ontologyTerms: {
-        '@id': 'has_ontology_term',
-        '@type': '@id',
-      },
-      cellTypeTerms: {
-        '@id': 'has_cell_type_term',
-        '@type': '@id',
-      },
-      thumbnail: {
-        '@id': 'has_thumbnail',
-      },
-    },
-    '@graph': [],
-  };
-  const donors = data['@graph'];
-
-  if (normalized !== '') {
-    for (const provider of normalized) {
-      if (provider.donors) {
-        const providerDonors = provider.donors.map((donor) => ({
-          consortium_name: provider.consortium_name,
-          provider_name: provider.provider_name,
-          provider_uuid: provider.provider_uuid,
-          ...donor,
-        }));
-        providerDonors.forEach((donor) =>
-          ensurePropertyOrder(donor, providerDonors)
-        );
-        providerDonors.forEach((donor) => donors.push(donor));
-      }
-    }
-  }
-
-  if (csv_normalized !== '') {
-    for (const data of csv_normalized) {
-      donors.push(data);
-    }
-  }
-
-  if (import_list_normalized !== '') {
-    for (const data of import_list_normalized) {
-      donors.push(data);
-    }
-  }
-
-  return data;
 }
